Fix lint violations from new code and split index write-back tests
